[MODIFY] Move Feed File to Social Directory
