--- app/src/main/java/com/cjwjsw/runningman/presentation/screen/feed/feedModel.kt
@@ -1,4 +1,0 @@
-package com.cjwjsw.runningman.presentation.screen.feed
-
-class feedModel {
-}
